+++ src/features/remembrance/members-view/component/MemorialIconGroup.tsx
