+++ src/features/remembrance/recipient-view/utils/getRecipientInfoItems.ts
