+++ src/pages/remembrance/MembersView.tsx
@@ -16,9 +16,9 @@ import {
 import TopArea from "@/shared/components/TopArea";
 import Description from "@/shared/components/Description";
 import { START_BEFORE, CHECK_ITEMS } from "@/shared/constant/members-view";
-import TributeArea from "@/features/members/component/TributeArea";
+import TributeArea from "@/features/remembrance/members-view/component/TributeArea";
 import CommentArea from "@/shared/components/comment/CommentArea";
-import HeavenLetterList from "@/features/members-view/component/HeavenLetterList";
+import HeavenLetterList from "@/features/remembrance/members-view/component/HeavenLetterList";
 import { withData } from "@/shared/utils/withData";
 
 export default function MembersView() {
